+++ account-kit/react/src/icons/timeout.tsx
@@ -1,188 +1,180 @@
-export const Timeout = () => {
+import { type SVGProps } from "react";
+import { useUiConfig } from "../hooks/useUiConfig.js";
+
+export const Timeout = ({
+  className,
+  ...props
+}: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) => {
+  const { illustrationStyle } = useUiConfig();
   return (
-    <svg
-      width="140"
-      height="141"
-      viewBox="0 0 140 141"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g clip-path="url(#clip0_55_11236)">
-        <circle cx="70" cy="70" r="70" fill="url(#paint0_linear_55_11236)" />
-        <g filter="url(#filter0_ddddd_55_11236)">
-          <rect
-            x="20.3846"
-            y="38"
-            width="99.2308"
-            height="102.308"
-            rx="8"
-            fill="white"
-          />
-        </g>
-        <circle cx="70" cy="71" r="18" fill="#F2A54A" />
-        <path
-          d="M70 62V72L75.5 76.5"
-          stroke="white"
-          stroke-width="3"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        />
-        <rect
-          x="48.6155"
-          y="105.308"
-          width="40"
-          height="6"
-          rx="2"
-          fill="url(#paint1_linear_55_11236)"
-        />
-        <rect
-          x="42.6155"
-          y="117.308"
-          width="54"
-          height="6"
-          rx="2"
-          fill="url(#paint2_linear_55_11236)"
-        />
-      </g>
-      <defs>
-        <filter
-          id="filter0_ddddd_55_11236"
-          x="-23.6154"
-          y="27"
-          width="215.231"
-          height="314.308"
-          filterUnits="userSpaceOnUse"
-          color-interpolation-filters="sRGB"
-        >
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0.901961 0 0 0 0 0.466667 0 0 0 0 0.176471 0 0 0 0.12 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="BackgroundImageFix"
-            result="effect1_dropShadow_55_11236"
-          />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dx="1" dy="9" />
-          <feGaussianBlur stdDeviation="10" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0.901961 0 0 0 0 0.466667 0 0 0 0 0.176471 0 0 0 0.12 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="effect1_dropShadow_55_11236"
-            result="effect2_dropShadow_55_11236"
-          />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dx="4" dy="36" />
-          <feGaussianBlur stdDeviation="18" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0.901961 0 0 0 0 0.466667 0 0 0 0 0.176471 0 0 0 0.1 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="effect2_dropShadow_55_11236"
-            result="effect3_dropShadow_55_11236"
-          />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dx="8" dy="81" />
-          <feGaussianBlur stdDeviation="24.5" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0.901961 0 0 0 0 0.466667 0 0 0 0 0.176471 0 0 0 0.06 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="effect3_dropShadow_55_11236"
-            result="effect4_dropShadow_55_11236"
-          />
-          <feColorMatrix
-            in="SourceAlpha"
-            type="matrix"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-            result="hardAlpha"
-          />
-          <feOffset dx="14" dy="143" />
-          <feGaussianBlur stdDeviation="29" />
-          <feColorMatrix
-            type="matrix"
-            values="0 0 0 0 0.901961 0 0 0 0 0.466667 0 0 0 0 0.176471 0 0 0 0.02 0"
-          />
-          <feBlend
-            mode="normal"
-            in2="effect4_dropShadow_55_11236"
-            result="effect5_dropShadow_55_11236"
-          />
-          <feBlend
-            mode="normal"
-            in="SourceGraphic"
-            in2="effect5_dropShadow_55_11236"
-            result="shape"
-          />
-        </filter>
-        <linearGradient
-          id="paint0_linear_55_11236"
-          x1="0"
-          y1="70"
-          x2="140"
-          y2="70"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#FAE0D1" />
-          <stop offset="1" stop-color="#FDF2CE" />
-        </linearGradient>
-        <linearGradient
-          id="paint1_linear_55_11236"
-          x1="46.1155"
-          y1="108.308"
-          x2="91.1155"
-          y2="108.308"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#F5F5F5" />
-          <stop offset="1" stop-color="#EBEBEB" stop-opacity="0.06" />
-        </linearGradient>
-        <linearGradient
-          id="paint2_linear_55_11236"
-          x1="39.2405"
-          y1="120.308"
-          x2="99.9905"
-          y2="120.308"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#F5F5F5" />
-          <stop offset="1" stop-color="#EBEBEB" stop-opacity="0.06" />
-        </linearGradient>
-        <clipPath id="clip0_55_11236">
-          <rect width="140" height="140.308" rx="70" fill="white" />
-        </clipPath>
-      </defs>
-    </svg>
+    <>
+      {illustrationStyle === "outline" && (
+        <TimedOutIconOutline className={className} {...props} />
+      )}
+      {illustrationStyle === "filled" && (
+        <TimedOutIconFilled className={className} {...props} />
+      )}
+      {illustrationStyle === "linear" && (
+        <TimedOutIconLinear className={className} {...props} />
+      )}
+      {illustrationStyle === "flat" && (
+        <TimedOutIconFlat className={className} {...props} />
+      )}
+    </>
   );
 };
+const TimedOutIconOutline = ({
+  className,
+  ...props
+}: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) => (
+  <svg
+    width="48"
+    height="48"
+    viewBox="0 0 48 48"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    className={className}
+    {...props}
+  >
+    <circle
+      cx="24.3428"
+      cy="27.5426"
+      r="18.0571"
+      stroke="#DC2626"
+      stroke-width="2.5"
+      className="stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+    <path
+      d="M24.3999 18.3999L24.3999 28.7999"
+      stroke="#DC2626"
+      stroke-width="2.5"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      className="stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+    <path
+      d="M19.3269 3.19995L29.3587 3.19995"
+      stroke="#DC2626"
+      stroke-width="2.5"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      className="stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+  </svg>
+);
+
+const TimedOutIconFilled = ({
+  className,
+  ...props
+}: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) => (
+  <svg
+    width="48"
+    height="48"
+    viewBox="0 0 48 48"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    className={className}
+    {...props}
+  >
+    <circle cx="24.3428" cy="27.5427" r="18.0571" fill="#EF4444" />
+    <path
+      d="M24.3999 18.4L24.3999 28.8"
+      stroke="white"
+      strokeWidth="2.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+    <path
+      d="M19.3269 3.20001L29.3587 3.20001"
+      stroke="#EF4444"
+      strokeWidth="2.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+);
+const TimedOutIconLinear = ({
+  className,
+  ...props
+}: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) => (
+  <svg
+    width="48"
+    height="48"
+    viewBox="0 0 48 48"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    className={className}
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M25.5 10.5324V29.5H23.5002L14.7756 12.536C17.3152 11.1731 20.2185 10.4 23.3024 10.4C24.0463 10.4 24.7797 10.445 25.5 10.5324Z"
+      fill="#FEE2E2"
+      className="fill-[#FEE2E2] dark:fill-[rgba(220,38,38,0.2)]"
+    />
+    <circle
+      cx="24.3428"
+      cy="27.5427"
+      r="18.0571"
+      stroke="#DC2626"
+      strokeWidth="2.5"
+      className="=stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+    <path
+      d="M24.3999 18.4L24.3999 28.8"
+      stroke="#DC2626"
+      strokeWidth="2.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className="=stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+    <path
+      d="M19.3269 3.20001L29.3587 3.20001"
+      stroke="#DC2626"
+      strokeWidth="2.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className="=stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+  </svg>
+);
+
+const TimedOutIconFlat = ({
+  className,
+  ...props
+}: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) => (
+  <svg
+    width="48"
+    height="48"
+    viewBox="0 0 48 48"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    className={className}
+    {...props}
+  >
+    <circle
+      cx="24.3428"
+      cy="27.5426"
+      r="18.0571"
+      fill="#FECACA"
+      className="fill-[#FEE2E2] dark:fill-[rgba(220,38,38,0.2)]"
+    />
+    <path
+      d="M24.3999 18.3999L24.3999 28.7999"
+      stroke="#DC2626"
+      strokeWidth="2.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className="stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+    <path
+      d="M19.3269 3.19995L29.3587 3.19995"
+      stroke="#DC2626"
+      strokeWidth="2.5"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className="stroke-[#DC2626] dark:stroke-[#EF4444]"
+    />
+  </svg>
+);